RoleType 추가 반영
