+++ src/app/(main)/users/register/termsText.js
